requirements.txt
Django==5.0.4
django-allauth==0.61.1
django-crispy-forms==2.1
django-embed-video==1.4.9
django-rest-swagger==2.2.0
django-simple-history==3.5.0
djangorestframework==3.15.1

coverage==7.4.0  # for testing
gunicorn==21.2.0
psycopg2-binary==2.9.9
pytest-django==4.7.0  # for testing
pytest==7.4.4  # for testing
wait-for-it==2.2.2
crispy_bootstrap5==2024.2
django_extensions==3.2.3
django-rest-passwordreset==1.4.1

# Rest apis
djangorestframework-simplejwt
django-filter
easy-thumbnails

drf-yasg
django-health-check
drf-nested-routers==0.94.1
django-cors-headers==4.3.1

python-decouple==3.8
pandas==2.2.2
openpyxl==3.1.5

django-import-export==4.1.1
drf-extra-fields==3.7.0